third_party/blink/web_tests/platform/generic/fast/dom/minor-dom-gc-expected.txt

PASS num2 is num1 + 4
PASS num3 is num1
PASS successfullyParsed is true

TEST COMPLETE
